+++ sdk/identity/azure-identity/tests/test_default_async.py
@@ -272,6 +272,23 @@ def get_credential_for_shared_cache_test(expected_refresh_token, expected_access
         return DefaultAzureCredential(_cache=cache, transport=transport, **exclude_other_credentials, **kwargs)
 
 
+def test_additionally_allowed_tenants():
+    """User-supplied additionally_allowed_tenants should be passed to applicable credentials"""
+
+    additionally_allowed_tenants = ["*"]
+    mocks = [Mock(), Mock(), Mock()]
+
+    with patch.multiple(DefaultAzureCredential.__module__, AzureCliCredential=mocks[0],
+        AzurePowerShellCredential=mocks[1], EnvironmentCredential=mocks[2]):
+
+        DefaultAzureCredential(additionally_allowed_tenants=additionally_allowed_tenants)
+
+    for mocked_credential in mocks:
+        _, kwargs = mocked_credential.call_args
+        assert "additionally_allowed_tenants" in kwargs
+        assert kwargs["additionally_allowed_tenants"] == additionally_allowed_tenants
+
+
 def test_unexpected_kwarg():
     """the credential shouldn't raise when given an unexpected keyword argument"""
     DefaultAzureCredential(foo=42)